move custom ckpt to own file
